frog-core/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==6.1.0
lxml==5.1.0
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.19.1
pypdf==4.2.0
python-docx==1.1.2
httpx==0.27.0
